move errors to libs
