+++ routes/index.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 var request = require('request');
-var HttpError = require('./errors').HttpError;
+var HttpError = require('../libs/errors').HttpError;
 var facebook = require('../libs/facebook');
 var winston = require('winston');
