+++ routes/jobs.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 var request = require('request');
-var HttpError = require('./errors').HttpError;
+var HttpError = require('../libs/errors').HttpError;
 var lodash = require('lodash');
 var winston = require('winston');
